Use 4 spaces for indentation instead of 2
I think 2 spaces for indentation is too little. By using 4 spaces it becomes more obvious that a piece of code is indented and it looks ugly when there's too many indentation levels -- which I also think should be avoided.

I would also be ok using tabs -- just pick one of those and stick with it.